mini_project_1/requirements.txt
#all the necessary environment 
streamlit
google-generativeai
python-dotenv